Complete Landing Page Assignment
I could do a lot differently here. I did not
want to spend more time on it because the time
would be better spent on the next sections of
the course.

Changes I would make would be to clean up the
CSS because i reuse a lot of things that could
be grouped together. I would also redo the HTML
to be cleaner, with less containers. This could
be simplified heavily